pkg/dxf: move node resource state to storage
